pyproject.toml
[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[project]
name = "recall"
version = "0.1.0"
description = "A tool to aggregate your digital activity, creating a perfect memory of your day."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client>=2.183.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "python-dotenv>=1.1.1",
    "python-gitlab>=6.4.0",
    "PyYAML>=6.0",
    "rich>=14.1.0",
    "slack-sdk>=3.36.0",
    "yaspin>=3.2.0",
]

[project.optional-dependencies]
dev = [
    "nox>=2025.5.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-watcher>=0.4.3",
    "pyfakefs>=5.4.0",
    "ruff>=0.13.3",
    "time-machine>=2.19.0",
]

[project.scripts]
recall = "recall.main:_main"

[tool.coverage.report]
show_missing = true

[tool.coverage.run]
branch = true

[tool.pytest.ini_options]
addopts = "--cov=recall --cov-report=term-missing"
pythonpath = ["src"]

[tool.ruff]
indent-width = 4
line-length = 88
target-version = "py39"

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
line-ending = "lf"
docstring-code-format = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D104",
    "D107",
    "FA100",
    "FA102",
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = [
    "ANN201",
    "PLR0913",
    "PLR2004",
    "S101",
    "S105",
    "S106",
    "SLF001",
]

"noxfile.py" = [
    "ANN201",
    "PLR0913",
    "PLR2004",
    "S101",
    "SLF001",
]
